requirements.txt
py-cord==2.0.0b4
mysql-connector-python
PyYAML